bindings: fix c binding 64 bit index type

+++ bindings/c/include/manifold/manifoldc.h
@@ -55,11 +55,11 @@ ManifoldMeshGL *manifold_meshgl_merge(void *mem, ManifoldMeshGL *m);
 
 ManifoldMeshGL64 *manifold_meshgl64(void *mem, double *vert_props,
                                     size_t n_verts, size_t n_props,
-                                    uint64_t *tri_verts, size_t n_tris);
+                                    size_t *tri_verts, size_t n_tris);
 
 ManifoldMeshGL64 *manifold_meshgl64_w_tangents(void *mem, double *vert_props,
                                                size_t n_verts, size_t n_props,
-                                               uint64_t *tri_verts,
+                                               size_t *tri_verts,
                                                size_t n_tris,
                                                double *halfedge_tangent);
 ManifoldMeshGL64 *manifold_get_meshgl64(void *mem, ManifoldManifold *m);
@@ -405,13 +405,13 @@ size_t manifold_meshgl64_run_transform_length(ManifoldMeshGL64 *m);
 size_t manifold_meshgl64_face_id_length(ManifoldMeshGL64 *m);
 size_t manifold_meshgl64_tangent_length(ManifoldMeshGL64 *m);
 double *manifold_meshgl64_vert_properties(void *mem, ManifoldMeshGL64 *m);
-uint64_t *manifold_meshgl64_tri_verts(void *mem, ManifoldMeshGL64 *m);
-uint64_t *manifold_meshgl64_merge_from_vert(void *mem, ManifoldMeshGL64 *m);
-uint64_t *manifold_meshgl64_merge_to_vert(void *mem, ManifoldMeshGL64 *m);
-uint64_t *manifold_meshgl64_run_index(void *mem, ManifoldMeshGL64 *m);
+size_t *manifold_meshgl64_tri_verts(void *mem, ManifoldMeshGL64 *m);
+size_t *manifold_meshgl64_merge_from_vert(void *mem, ManifoldMeshGL64 *m);
+size_t *manifold_meshgl64_merge_to_vert(void *mem, ManifoldMeshGL64 *m);
+size_t *manifold_meshgl64_run_index(void *mem, ManifoldMeshGL64 *m);
 uint32_t *manifold_meshgl64_run_original_id(void *mem, ManifoldMeshGL64 *m);
 double *manifold_meshgl64_run_transform(void *mem, ManifoldMeshGL64 *m);
-uint64_t *manifold_meshgl64_face_id(void *mem, ManifoldMeshGL64 *m);
+size_t *manifold_meshgl64_face_id(void *mem, ManifoldMeshGL64 *m);
 double *manifold_meshgl64_halfedge_tangent(void *mem, ManifoldMeshGL64 *m);
 
 // memory size
